Move getCalendar from routes/handlers to services
